refactor: decompose session.ts into focused modules
Extract surface management, memory recall/profile compilation, and
attachment resolution from the 2484-line session.ts into three dedicated
modules:

- session-surfaces.ts (510 lines): UI surface proxy resolver, undo,
  action handling, refresh, and completion summary
- session-memory.ts (197 lines): memory recall, dynamic profile,
  conflict gate evaluation
- session-attachments.ts (142 lines): host attachment approval,
  warning formatting, directive/tool-block resolution

The Session class retains its public API and delegates to the extracted
module functions via a SurfaceSessionContext interface pattern.
